refactor: 게임 불러오기 부분 컨트롤러 분리
